Break launcher into a separate binary
