+++ cmd/piped/BUILD.bazel
@@ -1,5 +1,7 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
-load("//bazel:image.bzl", "app_image")
+load("@io_bazel_rules_docker//go:image.bzl", "go_image")
+load("@io_bazel_rules_docker//container:container.bzl", "container_bundle", "container_image", "container_layer")
+load("@io_bazel_rules_docker//contrib:push-all.bzl", "docker_push")
 
 go_library(
     name = "go_default_library",
@@ -7,7 +9,6 @@ go_library(
     importpath = "github.com/pipe-cd/pipe/cmd/piped",
     visibility = ["//visibility:private"],
     deps = [
-        "//pkg/app/piped/cmd/launcher:go_default_library",
         "//pkg/app/piped/cmd/piped:go_default_library",
         "//pkg/cli:go_default_library",
     ],
@@ -19,18 +20,77 @@ go_binary(
     visibility = ["//visibility:public"],
 )
 
-app_image(
-    name = "piped_app",
+# Piped app
+
+container_layer(
+    name = "launcher",
+    tars = [
+        "//cmd/launcher:launcher_image",
+    ],
+)
+
+go_image(
+    name = "piped_image",
     base = "@piped-base//image",
     binary = ":piped",
-    repository = "piped",
+)
+
+container_image(
+    name = "piped_app_image",
+    base = ":piped_image",
+    layers = [
+        ":launcher",
+    ],
+    symlinks = {
+        "/piped": "/app/cmd/piped/piped",
+        "/launcher": "/app/cmd/launcher/launcher",
+    },
     visibility = ["//visibility:public"],
 )
 
-app_image(
-    name = "piped_okd_app",
+container_bundle(
+    name = "piped_app_bundle",
+    images = {
+        "$(DOCKER_REGISTRY)/piped:{STABLE_VERSION}": ":piped_app_image",
+    },
+)
+
+docker_push(
+    name = "piped_app_push",
+    bundle = ":piped_app_bundle",
+    visibility = ["//visibility:public"],
+)
+
+# Piped app for OKD
+
+go_image(
+    name = "piped_okd_image",
     base = "@piped-base-okd//image",
     binary = ":piped",
-    repository = "piped-okd",
+)
+
+container_image(
+    name = "piped_okd_app_image",
+    base = ":piped_okd_image",
+    layers = [
+        ":launcher",
+    ],
+    symlinks = {
+        "/piped": "/app/cmd/piped/piped",
+        "/launcher": "/app/cmd/launcher/launcher",
+    },
+    visibility = ["//visibility:public"],
+)
+
+container_bundle(
+    name = "piped_okd_app_bundle",
+    images = {
+        "$(DOCKER_REGISTRY)/piped-okd:{STABLE_VERSION}": ":piped_okd_app_image",
+    },
+)
+
+docker_push(
+    name = "piped_okd_app_push",
+    bundle = ":piped_okd_app_bundle",
     visibility = ["//visibility:public"],
 )
