+++ pkg/app/launcher/cmd/launcher/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "binary.go",
         "launcher.go",
     ],
-    importpath = "github.com/pipe-cd/pipe/pkg/app/piped/cmd/launcher",
+    importpath = "github.com/pipe-cd/pipe/pkg/app/launcher/cmd/launcher",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/app/api/service/pipedservice:go_default_library",
